Remove version information consts from libxfce4gui
The version information is only useful in C sources, there is no way to
do #if LIBFOO_CHECK_VERSION (0,1,2) in Vala. So, out!

Update the test program to flush the RC data before opening it again for
reading. In fact, when creating the new instance, it is created inside a
temp var, then the original var is closed (=flushed too btw), and
finally the temp var is assigned to the original var. So, flush!