More updates on exo-0.3 bindings
- Add a test file. It's mostly incomplete, it doesn't test any of the
widgets as they were very good introspected, the MountPoint class is
pretty annoying to use and I simply suggest "use GIO", and I didn't
succeed to use String.dup_strftime properly with a struct tm. The CCode
made use of 'tm' and not 'struct tm', might be a bug in the Posix
bindings... but anyway there are date and time functions in GLib.
- Fixed Binding classes.
- Fixed Md5Digest functions, moved str_get_md5_str to new class Md5.
- Fixed String class and some weak references.