Make the panel plugin example a bit more complete
The plugin needs to call Gtk.main_quit() when it gets free'd or
destroy'd. I settled for the destroy signal cause it separates a bit the
logic of what do do on free or destroy. This is new, and if you don't do
this call, the wrapper process will still be around after you remove the
plugin from the panel.

Connected a few signal to populate the example and made the button an
action widget so it can be right clicked.